Move decorator to utils.deprecation
