+++ qiskit/providers/ibmq/utils/__init__.py
@@ -14,5 +14,6 @@
 
 """Utilities related to the IBMQ Provider."""
 
+from .deprecation import deprecated
 from .qobj_utils import update_qobj_config
